requirements.txt
black
black[jupyter]
pandas
numpy
Jinja2
urllib3
MLB-StatsAPI==1.7
pybaseball
requests
beautifulsoup4
multiprocess
ipython
jupyterlab
patsy
scikit-learn
tensorflow